Remove zero padding when hashing files smaller than PREHASH_SIZE
On the program's second pass if a file is smaller than PREHASH_SIZE
the program will write data into the start of the buffer and leave
the end of the buffer with its previous values. Then the program will
hash the entire buffer regardless if some of the buffer does not
contain file data.

This extra data at the end of the file is currently guarenteed to
be zero because all the files are hashed from smallest to largest
because they are stored in a BTreeMap keyed by file size. That means
new file data will always overwrite previous data in the buffer
with the rest of the buffer initalized to zero.